pyproject.toml
[tool.poetry]
name = "cloud-custodian-policy"
version = "0.1.0"
description = "Cloud Governance Platform Design"
authors = ["Kent <kentnsw@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
c7n = "^0.9.32"
c7n-org = "^0.6.31"
c7n-mailer = "^0.6.31"


[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
black = "^23.11.0"
moto = "^4.2.9"
freezegun = "^1.2.2"
boto3-stubs = {version = "1.28.58", extras = ["cloudformation", "dynamodb", "ec2", "lambda", "rds", "s3", "sqs"]}
loguru = "^0.7.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
